refactor(database): toolbar and row container
